pyproject.toml
[project]
name = "ai-news-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.12",
    "groq>=0.4.0",
    "fastapi>=0.115.5",
    "uvicorn>=0.32.0",
    "pymongo>=4.8.0",
    "python-jose>=3.3.0",
    "bcrypt>=4.1.2",
    "razorpay>=1.4.2",
    "google-generativeai>=0.8.3",
    "html-to-markdown>=2.7.1",
    "markdown>=3.7.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.44",
    "youtube-transcript-api>=1.2.3",
    "apscheduler>=3.10.4",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
